Fix a crash vulnerability caused by invalid user callback
It is possible for users to send unexpected callbacks to crash the whole module. This
can be done by modifying the href strings at the client side and then clicking on
the altered links.

Ensure that the module does not crash because of this by safely calling the handlers
via protected call (pcall).  Print the error in case the crash was indeed caused by a
coding mistake.